+++ lib/gyoku/xml_value.rb
@@ -8,15 +8,19 @@ module Gyoku
       # xs:date format
       XS_DATE_FORMAT = "%Y-%m-%d"
 
+      XS_TIME_FORMAT = "%H:%M:%S"
+
       # xs:dateTime format
       XS_DATETIME_FORMAT = "%Y-%m-%dT%H:%M:%S%Z"
 
       # Converts a given +object+ to an XML value.
       def create(object, escape_xml = true)
-        if Date === object
-          object.strftime XS_DATE_FORMAT
+        if Time === object
+          object.strftime XS_TIME_FORMAT
         elsif DateTime === object
           object.strftime XS_DATETIME_FORMAT
+        elsif Date === object
+          object.strftime XS_DATE_FORMAT
         elsif String === object
           escape_xml ? CGI.escapeHTML(object) : object
         elsif object.respond_to?(:to_datetime)